refactor adapter files for cleaner source. In detail, move view holder inside adapter.
