+++ app/src/main/java/hcmute/edu/vn/noicamheo/adapter/ContactAdapter.java
@@ -5,7 +5,10 @@ import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ImageView;
+import android.widget.LinearLayout;
 import android.widget.RelativeLayout;
+import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
@@ -97,5 +100,43 @@ public class ContactAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder
     @Override
     public int getItemCount() {
         return contacts.size();
+    }
+
+    // Represent a contact element in contact list
+    private class ContactViewHolder extends RecyclerView.ViewHolder {
+        RelativeLayout mainLayout;
+        TextView textViewFullName;
+        RelativeLayout openPanel;
+        TextView textViewPhone;
+        ImageView imageViewCall;
+        ImageView imageViewMessage;
+        LinearLayout divider;
+        RelativeLayout.LayoutParams params;
+
+
+        public ContactViewHolder(@NonNull View itemView) {
+            super(itemView);
+            mainLayout = itemView.findViewById(R.id.itemContact);
+            textViewFullName = itemView.findViewById(R.id.textViewFullName);
+            openPanel = itemView.findViewById(R.id.relativeLayoutOpeningPanel);
+            textViewPhone = itemView.findViewById(R.id.textViewPhone);
+            imageViewCall = itemView.findViewById(R.id.imageViewCall);
+            imageViewMessage = itemView.findViewById(R.id.imageViewMessage);
+            divider = itemView.findViewById(R.id.contactItemDivider);
+            params = (RelativeLayout.LayoutParams) divider.getLayoutParams();
+        }
+
+        public RelativeLayout.LayoutParams getDividerParam() {
+            return params;
+        }
+    }
+
+    // Represent a contact header (first letter name group)
+    private class ContactHeaderViewHolder extends RecyclerView.ViewHolder {
+        TextView textViewHeader;
+        public ContactHeaderViewHolder(@NonNull View itemView) {
+            super(itemView);
+            textViewHeader = itemView.findViewById(R.id.textViewHeader);
+        }
     }
 }
