--- app/src/main/java/hcmute/edu/vn/noicamheo/adapter/ContactHeaderViewHolder.java
@@ -1,17 +1,0 @@
-package hcmute.edu.vn.noicamheo.adapter;
-
-import android.view.View;
-import android.widget.TextView;
-
-import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.RecyclerView;
-
-import hcmute.edu.vn.noicamheo.R;
-
-public class ContactHeaderViewHolder extends RecyclerView.ViewHolder {
-    TextView textViewHeader;
-    public ContactHeaderViewHolder(@NonNull View itemView) {
-        super(itemView);
-        textViewHeader = itemView.findViewById(R.id.textViewHeader);
-    }
-}
